GEODE-2484: Replace ACE Map with synchronized unordered_map (#401)
* Adds an synchronized_map wrapper template to synchronize a map with a
  mutex.
* Replaces unsynchronized ACE Map with std::unordered_map.
* Replaces ACE synchronized Set with synchronized std::set
* Replaces ACE_Semaphore with std::mutex in ThinClientRegion
* Replaces ACE_Semaphore with std::condition_variable in ClientMetaDataService
* Replaces ACE_Task with std::thread in ClientMetaDataService
* Replaces ACE_RW_Thread_Mutex with boost::shared_mutex in ClientMetaDataService
* Replaces use of synchronize bounded queue.
 * Was not using the bounded aspect of the Queue type.
 * Converted to using std::deque directly.
 * Refactored out coalesce function.
* Replaces ACE_Task with std::thread in EvictionController
* Replace ACE_RW_Thread_Mutex with boost::shared_mutex in EvictionController
* Replace specialized IntQueue with std::queue in EvictionController
 * Use condition variables to avoid spurious wakeups.
* Replace raw pointer allocation with std::unique_ptr
* Replace specialized IntQueue with std::deque
* Replace ACE_Task with std::thread.
 * Adds new wrapper based on std::thread. 
* Cleanup includes for fast compilation.
* Use std::unique_ptr to manage destruct.
* Adds thread names to MacOS and Windows threads.

+++ cppcache/src/CacheFactory.cpp
@@ -108,9 +108,8 @@ Cache CacheFactory::create() const {
       std::bind(VersionTag::createDeserializable, memberListForVersionStamp));
 
   serializationRegistry->addDataSerializableFixedIdType(
-      static_cast<int64_t>(DSFid::DiskVersionTag),
-      std::bind(DiskVersionTag::createDeserializable,
-                memberListForVersionStamp));
+      DSFid::DiskVersionTag, std::bind(DiskVersionTag::createDeserializable,
+                                       memberListForVersionStamp));
 
   serializationRegistry->setPdxTypeHandler(new PdxTypeHandler());
   serializationRegistry->setDataSerializableHandler(
